requirements.txt
python-dotenv
assemblyai
openai